+++ includes/exec.js
@@ -1,0 +1,13 @@
+const exec = require("child_process").exec;
+
+module.exports = function(cmd) {
+  exec(cmd, function(error, stdout, stderr) {
+    if (error) {
+      console.error(`exec error: ${error}`);
+      return;
+    }
+    if (stderr) {
+      console.error(`stderr: ${stderr}`);
+    }
+  });
+};