+++ commands/lint.js
@@ -2,8 +2,15 @@ const exec = require("../includes/exec");
 
 module.exports = function() {
   exec("prettier");
-  exec("eslint --fix --ext .js,.jsx,.ts,.tsx src doc/src");
-  exec(
-    "stylelint --fix src/**/*.css src/**/*.scss doc/src/**/*.css doc/src/**/*.scss"
-  );
+  const scriptExts = [".js", ".jsx", ".ts", ".tsx"].join(",");
+  exec("eslint --fix --ext " + scriptExts + " .");
+  const styleFiles = [
+    "*.css",
+    "*.scss",
+    "src/**/*.css",
+    "src/**/*.scss",
+    "doc/src/**/*.css",
+    "doc/src/**/*.scss"
+  ].join(" ");
+  exec("stylelint --fix " + styleFiles);
 };